Changed Capacity for local PVs to use FsInfo capacity
Earlier version used "available" from FsInfo
New version uses "capacity" from FsInfo
I also changed uint64 to int64

+++ local-volume/provisioner/pkg/discovery/discovery_test.go
@@ -324,7 +324,7 @@ func verifyCapacity(t *testing.T, createdPV *v1.PersistentVolume, expectedPV *te
 	if !ok {
 		t.Errorf("Unable to convert resource storage into int64")
 	}
-	if uint64(capacityInt) != expectedPV.capacity {
+	if capacityInt != expectedPV.capacity {
 		t.Errorf("Expected capacity %d, got %d", expectedPV.capacity, capacityInt)
 	}
 }
@@ -333,7 +333,7 @@ func verifyCapacity(t *testing.T, createdPV *v1.PersistentVolume, expectedPV *te
 type testPVInfo struct {
 	pvName       string
 	path         string
-	capacity     uint64
+	capacity     int64
 	storageClass string
 }
